requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
appnope=0.1.4=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.6.2=h8857fd0_0
certifi=2024.6.2=pypi_0
cffi=1.16.0=pypi_0
click=8.1.7=pypi_0
comm=0.2.2=pyhd8ed1ab_0
cryptography=42.0.8=pypi_0
debugpy=1.8.1=py311hdd0406b_0
decorator=5.1.1=pyhd8ed1ab_0
dnspython=2.6.1=pypi_0
email-validator=2.2.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.4=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=7.2.0=pyha770c72_0
importlib_metadata=7.2.0=hd8ed1ab_0
iniconfig=2.0.0=pypi_0
ipykernel=6.29.4=pyh57ce528_0
ipython=8.25.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py311h6eed73b_0
krb5=1.21.2=hb884880_0
libcxx=17.0.6=h88467a6_0
libedit=3.1.20191231=h0678c8f_2
libffi=3.4.4=hecd8cb5_1
libsodium=1.0.18=hbcb3906_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
openssl=3.3.1=h87427d6_0
orjson=3.10.5=pypi_0
packaging=24.1=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pip=24.0=py311hecd8cb5_0
platformdirs=4.2.2=pyhd8ed1ab_0
pluggy=1.5.0=pypi_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=5.9.8=py311he705e18_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.22=pypi_0
pydantic=2.7.4=pypi_0
pydantic-core=2.18.4=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyjwt=2.8.0=pypi_0
pytest=8.2.2=pypi_0
python=3.11.9=hf27a42d_0
python-dateutil=2.9.0=pyhd8ed1ab_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
python_abi=3.11=2_cp311
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=py311h89e2aaa_0
readline=8.2=hca72f7f_0
rich=13.7.1=pypi_0
setuptools=69.5.1=py311hecd8cb5_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h6c40b1e_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.37.2=pypi_0
tk=8.6.14=h4d00af3_0
tornado=6.4.1=py311h72ae277_0
traitlets=5.14.3=pyhd8ed1ab_0
typer=0.12.3=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h04d1e81_0
ujson=5.10.0=pypi_0
uvicorn=0.30.1=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.22.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
websockets=12.0=pypi_0
wheel=0.43.0=py311hecd8cb5_0
xz=5.4.6=h6c40b1e_1
zeromq=4.3.5=hde137ed_4
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.2.13=h4b97444_1
